Updated the way user_account is initialised, with an init function
I created an init function that first runs the superclass initialisation,
and then continues with initialising user_account to the value of a getter
function. The getter function currently returns the dummy value defined in
the test data, but will be updated in the future, when we decide how to
store the trusted account number.

Signed-off-by: Deniz Genc <25902330+denizgenc@users.noreply.github.com>