Added extra detail to passing test cases
Previously, the pass case had a single IAM User ARN within a list, which
does not happen. If a principal trusts a single entity, they will not be
inside a list.

Example of previous test data:
{"AWS" : ["some:user:arn"]}

Example of new, correct test data:
{"AWS" : "some:user:arn"}

Currently the criterion does not pass the unit test, so I will be
changing it to do so.