Test data for access denied error
My intial additions actually caused get_data to complain - the
access_denied versioning data was a string, instead of a dict like
get_data expects. Because of that, I decided to put it into a dict and
put the exception string as the value of a key I called
"csw_access_denied_error".
This will require changes in the client and check code, but it will mean
that the check code will become marginally more simple: instead of
checking whether the versioning data is a dict or a string, we can
assume it will always be a dict and continue from there.