Set annotation to an empty string when evaluation is compliant
This change was prompted by what seemed like a bug in CloudWatch:
resources with an evaluation of COMPLIANT still had an annotation that
said "This S3 bucket does not have versioning enabled."
We eventually concluded that if a NON-COMPLIANT resource was evaluated
before a COMPLIANT one, the annotation was not getting overwritten,
carrying over the previous when building the evaluation for the
COMPLIANT resource.
I don't know how prevalent this error is for other checks, but we have
not been testing for it in general, so it may have been slipping
through.